changed bit length param back to uint8_t

+++ core/include/bitwise_utils.h
@@ -7,15 +7,15 @@ typedef uint64_t BitMatrix64[64];
 
 // bit vector operations
 
-static inline void bv32_set_bit(uint32_t* vector, uint32_t k) {
+static inline void bv32_set_bit(uint32_t* vector, uint8_t k) {
     vector[k >> 5] |= 1 << (k & 0x1F);
 }
 
-static inline void bv32_clear_bit(uint32_t* vector, uint32_t k) {
+static inline void bv32_clear_bit(uint32_t* vector, uint8_t k) {
     vector[k >> 5] &= ~(1 << (k & 0x1F));
 }
 
-static inline bool bv32_get_bit(const uint32_t* vector, uint32_t k) {
+static inline bool bv32_get_bit(const uint32_t* vector, uint8_t k) {
     return (vector[k >> 5] >> (k & 0x1F)) & 1;
 }
 
@@ -50,15 +50,15 @@ void bm64_transpose(BitMatrix64 matrix);
 
 // bitwise operations
 
-static inline uint32_t mask_bits(uint32_t n) {
+static inline uint32_t mask_bits(uint8_t n) {
     return ~0u >> (32 - n);
 }
 
-static inline uint32_t invert_bits(uint32_t x, uint32_t n) {
+static inline uint32_t invert_bits(uint32_t x, uint8_t n) {
     return ~x & mask_bits(n);
 }
 
-uint32_t reverse_bits(uint32_t x, uint32_t n);
+uint32_t reverse_bits(uint32_t x, uint8_t n);
 
 #define count_trailing_zeros(X) ((X) ? perfect_log2((X) & -(X)) : sizeof(X) * 8)
 #define perfect_log2(X) (sizeof(X) <= sizeof(uint32_t) ? perfect_log2_32(X) : perfect_log2_64(X))
